feat(http): change test location
